Fix goroutine leaks of updateServsrDate (#1241).
* updateServerDate's goroutine makes available to catch s.done closing
